@@ -2,17 +2,17 @@ import React from "react";
 import { Link } from "react-router-dom";
 
 
-const ProductCard = (props) => {
-  const { id, productCode, productDescription } = props.product;
+const DriverCard = (props) => {
+  const { id, driverCode, driverDescription } = props.driver;
   return (
     <div className="item">
       
       <div className="content">
         <Link
-          to={{ pathname: `/product/${id}`, state: { product: props.product } }}
+          to={{ pathname: `/driver/${id}`, state: { driver: props.driver } }}
         >
-          <div className="header">{productCode}</div>
-          <div>{productDescription}</div>
+          <div className="header">{driverCode}</div>
+          <div>{driverDescription}</div>
         </Link>
       </div>
       <i
@@ -20,7 +20,7 @@ const ProductCard = (props) => {
         style={{ color: "red", marginTop: "7px", marginLeft: "10px" }}
         onClick={() => props.clickHandler(id)} 
       ></i>
-      <Link to={{ pathname: `/edit`, state: { product: props.product } }}>
+      <Link to={{ pathname: `/edit`, state: { driver: props.driver } }}>
         <i
           className="edit alternate outline icon"
           style={{ color: "blue", marginTop: "7px" }}
@@ -30,4 +30,4 @@ const ProductCard = (props) => {
   );
 };
 
-export default ProductCard;
+export default DriverCard;
